\name{adjacent.dispersal}
\alias{adjacent.dispersal}
\title{A dispersal function that only has 3 values, one for local, one for adjacent
cells, and zero for others.  t1 and t2 are threshold distances.}
\usage{
  adjacent.dispersal(distance, t1, t2, local, adjacent)
}
\description{
  A dispersal function that only has 3 values, one for
  local, one for adjacent cells, and zero for others.  t1
  and t2 are threshold distances.
}

