% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DensityDependence.R
\name{DensityDependence}
\alias{DensityDependence}
\title{Produced multiplier for new recruits given population and space vectors}
\usage{
DensityDependence(pop, space)
}
\description{
Produced multiplier for new recruits given population and space vectors
}

