\name{FitLattice}
\alias{FitLattice}
\title{Create a hex grid around a set of site locations}
\usage{
  FitLattice(locations, plot.area = 500, buffer = 0.1)
}
\arguments{
  \item{locations}{a data frame or matrix of locations.
  The last two column must be the X and Y locations of the
  sites}

  \item{plot.area}{the area (in sq. meters) of the circular
  plot locations.  Used to calculate the grid size so hex
  cells are of the same area}

  \item{buffer}{the amount of buffer space around the
  edgemost points to put in the grid.  As a fraction of the
  total distance between the extreme points.}
}
\value{
  An object of the class "lattice" which is a matrix
  generated by \code{\link{MakeHexLattice}} of site
  locations.  It has the additional attribute \code{nn}
  which is the \code{locations} parameter with an
  additional column, \code{lattice.nn} denoting the
  nearest-neighbor points on the lattice to each of the
  plot locations.
}
\description{
  Creates a hex lattice of points overlain on a series of
  plot locations, matching each real plot location with its
  nearest neighbor grid point. The origin is adjusted to
  minimize the sum squared error difference between actual
  and grid plot locations.
}

