\name{MakeDispMatrix}
\alias{MakeDispMatrix}
\title{Generate a dispersal matrix from locations.  The matrix calculates pairwise
distances between locations weighted by the dispersal kernel of each class}
\usage{
  MakeDispMatrix(parms.df, locations, parms.obj)
}
\description{
  Generate a dispersal matrix from locations.  The matrix
  calculates pairwise distances between locations weighted
  by the dispersal kernel of each class
}

