\name{MakeGrid}
\alias{MakeGrid}
\title{Generate a lattice of equally spaced locations and coordinates}
\usage{
  MakeGrid(nx = NULL, ny = NULL, dist = 1, type = "square",
    area = NULL, origin = c(0, 0), around = NULL,
    xy = c("Easting", "Northing"))
}
\description{
  Generate a lattice of equally spaced locations and
  coordinates
}

