\name{MakeHexLattice}
\alias{MakeHexLattice}
\title{Generate a lattice of points equally spaced in the centers of a hexagonal
lattice}
\usage{
  MakeHexLattice(nx, ny, dist = 1, origin = c(0, 0))
}
\description{
  Generate a lattice of points equally spaced in the
  centers of a hexagonal lattice
}

