% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MakeLattice.R
\name{MakeLattice}
\alias{MakeLattice}
\title{Generate a lattice of equally spaced locations and coordinates}
\usage{
MakeLattice(nx, ny, dist = 1)
}
\description{
Generate a lattice of equally spaced locations and coordinates
}

