% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SODModel.R
\name{SODModel}
\alias{SODModel}
\title{Runs the disease model.  Outputs a large matrix of population by species, ageclass, location}
\usage{
SODModel(parms.df, locations, time.steps, init, df.out = TRUE,
  verbose = interactive())
}
\description{
Runs the disease model.  Outputs a large matrix of population by species, ageclass, location
}

