\name{SODModelrun}
\alias{SODModelrun}
\title{Runs the disease model.  Outputs a large matrix of population by species,
sizeclass, location}
\usage{
  SODModelrun(parms, times, locations, init, lambda.ex, K,
    stochastic.e, stochastic.d, verbose, run = NULL)
}
\description{
  Runs the disease model.  Outputs a large matrix of
  population by species, sizeclass, location
}

