\name{SODSims}
\alias{SODSims}
\title{Run many simulations}
\usage{
  SODSims(inits, .parallel = FALSE, verbose = TRUE, ...)
}
\arguments{
  \item{nsims}{The number of simulations to perform}

  \item{parallel}{Run the simulation in parallel or not?}

  \item{...}{arguments to be passed to SODModel}

  \item{seed}{numeric. If used, sets the seed for
  simulations}
}
\description{
  Run many simulations
}

