\name{adjacent.dispersal}
\alias{adjacent.dispersal}
\title{A dispersal function that only has 3 values, one for local, one for adjacent
cells, and zero for others.  Requires distances between cells to = 1}
\usage{
  adjacent.dispersal(distance, local, adjacent)
}
\description{
  A dispersal function that only has 3 values, one for
  local, one for adjacent cells, and zero for others.
  Requires distances between cells to = 1
}

