% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SODModel.R
\name{MakeDispMatrix}
\alias{MakeDispMatrix}
\title{Generate a dispersal matrix from locations.}
\usage{
MakeDispMatrix(parms.df, locations, parms.obj)
}
\description{
Generate a dispersal matrix from locations.
}

