\name{SODModel}
\alias{SODModel}
\title{Run the SOD model}
\usage{
  SODModel(parms, times, locations, init, lambda.ex = 0,
    stochastic.e = NULL, stochastic.d = FALSE,
    verbose = interactive(), .parallel = FALSE)
}
\arguments{
  \item{.parallel}{Run the simulation in parallel or not?}

  \item{...}{arguments to be passed to SODModel}
}
\description{
  Run the SOD model
}

