\name{SODModel}
\alias{SODModel}
\title{Runs the disease model.  Outputs a large matrix of population by species, sizeclass, location}
\usage{
  SODModel(parms.df, locations, time.steps, init,
    lambda.ex = 0, df.out = TRUE, verbose = interactive(),
    stochastic.d = FALSE, stochastic.e = FALSE)
}
\description{
  Runs the disease model.  Outputs a large matrix of
  population by species, sizeclass, location
}

