\name{SODModel}
\alias{SODModel}
\title{Runs the disease model.  Outputs a large matrix of population by species, sizeclass, location}
\usage{
  SODModel(parms.df, locations, time.steps, init,
    df.out = TRUE, verbose = interactive())
}
\description{
  Runs the disease model.  Outputs a large matrix of
  population by species, sizeclass, location
}

